created all entities in database, added create course and delete course endpoints

+++ api/Services/Impls/GroupService.cs
@@ -1,7 +1,10 @@
 ﻿using api.Exceptions;
 using api.Mappers;
+using api.Models.CampusCourse;
 using api.Models.CampusGroup;
 using api.Services.Interfaces;
+using Microsoft.EntityFrameworkCore;
+using Newtonsoft.Json.Linq;
 
 namespace api.Services.Impls
 {
@@ -29,13 +32,18 @@ namespace api.Services.Impls
                 throw new UnauthorizedException(ErrorConstants.UnauthorizedError);
             }
             var user = await _accountService.GetUserByToken(token);
+            var role = await _db.Roles.FirstOrDefaultAsync(r => r.UserId == user.Id);
+            if (role == null || !role.IsAdmin)
+            {
+                throw new ForbiddenException(ErrorConstants.ForbiddenError);
+            }
             var group = GroupMapper.MapFromCreateCampusGroupModelToCampusGroup(createCampusGroupModel);
             await _db.CampusGroups.AddAsync(group);
             await _db.SaveChangesAsync();
             return GroupMapper.MapFromCampusGroupToCampusGroupModel(group);
         }
 
-        public async Task<List<CampusGroupModel>> GetCampusGroup(string token)
+        public async Task<List<CampusGroupModel>> GetCampusGroups(string token)
         {
             if (string.IsNullOrWhiteSpace(token))
             {
@@ -52,6 +60,12 @@ namespace api.Services.Impls
 
         public async Task<CampusGroupModel> EditCampusCourse(EditCampusGroupModel editCampusGroupModel, Guid groupId, string token, string userId)
         {
+            var user = await _accountService.GetUserByToken(token);
+            var role = await _db.Roles.FirstOrDefaultAsync(r => r.UserId == user.Id);
+            if (role == null || !role.IsAdmin)
+            {
+                throw new ForbiddenException(ErrorConstants.ForbiddenError);
+            }
             var group = GroupMapper.MapFromEditCampusGroupModelToCampusGroup(editCampusGroupModel, groupId);
             _db.CampusGroups.Update(group);
             await _db.SaveChangesAsync();
@@ -63,9 +77,28 @@ namespace api.Services.Impls
             var user = await _accountService.GetUserById(userId);
             var group = await _db.CampusGroups.FindAsync(id);
             if (group == null) throw new NotFoundException(ErrorConstants.NotFoundGroupError);
-            // if (user.Id != comment.AuthorId) throw new CustomException("You are not the author of this comment", 403);
+            var role = await _db.Roles.FirstOrDefaultAsync(r => r.UserId == user.Id);
+            if (role == null || !role.IsAdmin)
+            {
+                throw new ForbiddenException(ErrorConstants.ForbiddenError);
+            }
             _db.Remove(group);
             await _db.SaveChangesAsync();
+        }
+
+        public async Task<List<CampusCoursePreviewModel>> GetCampusGroup(Guid groupId, string userId)
+        {
+            var user = await _accountService.GetUserById(userId);
+            var group = await _db.CampusGroups.FindAsync(groupId);
+            if (group == null) throw new NotFoundException(ErrorConstants.NotFoundGroupError);
+            var role = await _db.Roles.FirstOrDefaultAsync(r => r.UserId == user.Id);
+            if (role == null || !role.IsAdmin)
+            {
+                throw new ForbiddenException(ErrorConstants.ForbiddenError);
+            }
+            var course = await _db.Courses.FirstOrDefaultAsync(c => c.CampusGroupId == groupId);
+            var courses = _db.Courses.Select(course => CourseMapper.MapFromCampusCourseToCampusCoursePreviewModel(course)).ToList();
+            return courses;
         }
     }
 }
